Modified template table of contents
Removed secondary level content as it'd likely be too burdensome to manage regularly.  (Each link would have to be updated to the name of each item of secondary level content.)  

The primary level content in the table of contents won't change - the names will stay the same - which won't require editing the table of contents every time a pattern is authored.